requirements.txt
web3>=6.0.0
eth-account>=0.9.0
python-dotenv>=1.0.0
termcolor>=2.3.0
pyfiglet>=0.8.0
colorama>=0.4.6
